src/python/grpc-server/pyproject.toml
[tool.poetry]
name = "grpc-server"
version = "0.1.0"
description = ""
authors = ["pecolynx <pecolynx@gmail.com>"]
readme = "README.md"
packages = [{include = "grpc_server"}]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.59.0"
protobuf = "^4.24.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
